// angular-front/src/app/model/profissional.model.ts
export class Profissional {

    idProfissional: number;
    nome: string;

    logradouro: string;
    numero: string;
    cep: number;
    bairro: string;
    cidade: string;
    uf: string;
    pais: string;
    complemento: string;
    
    idEstabProfissional: number;
    nomeEstabProfissional: string;

}